+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/integration/recipemod/jei/RefinedStorageJeiModPlugin.java
@@ -56,6 +56,10 @@ public class RefinedStorageJeiModPlugin implements IModPlugin {
             AbstractGridScreen.class,
             new GridGuiContainerHandler(converter, registration.getJeiHelpers().getIngredientManager())
         );
+        registration.addGenericGuiContainerHandler(
+            AbstractBaseScreen.class,
+            new ExclusionZonesGuiContainerHandler()
+        );
     }
 
     @Nullable